feat(locale): add locale emulation + tests
