+++ emulator-plugins/shared/test/chrome.test.ts
@@ -91,7 +91,7 @@ async function createPage() {
     {
       proxyPassword: '',
       platform: 'win32',
-      acceptLanguage: 'en',
+      locale: 'en',
       userAgent: 'Chrome Test',
       viewport: {
         screenHeight: 900,
