+++ emulators/lib/EmulatorPlugin.ts
@@ -10,11 +10,16 @@ export default abstract class EmulatorPlugin {
   public abstract engineExecutablePath: string;
   public abstract engine: { browser: string; revision: string };
   public abstract delegate: IHttpRequestModifierDelegate;
+  public locale = 'en-US,en';
   protected userProfile: IUserProfile;
 
   public abstract async generatePageOverrides(): Promise<IPageOverride[]>;
 
   public setUserProfile(userProfile: IUserProfile) {
     this.userProfile = userProfile;
+  }
+
+  public setLocale(locale: string) {
+    this.locale = locale;
   }
 }
