+++ full-client/package.json
@@ -11,6 +11,7 @@
   "devDependencies": {
     "@secret-agent/commons": "1.0.0-alpha.20",
     "@secret-agent/emulate-chrome-80": "1.0.0-alpha.20",
+    "@secret-agent/emulators": "1.0.0-alpha.20",
     "@secret-agent/mitm": "1.0.0-alpha.20",
     "@secret-agent/testing": "1.0.0-alpha.20",
     "@types/node": "^12.7.11",
