+++ puppet-chrome/lib/Page.ts
@@ -21,6 +21,7 @@ import { IRegisteredEventListener, TypedEventEmitter } from '@secret-agent/commo
 import { createPromise } from '@secret-agent/commons/utils';
 import { IBoundLog } from '@secret-agent/commons/Logger';
 import IViewport from '@secret-agent/core-interfaces/IViewport';
+import { CanceledPromiseError } from '@secret-agent/commons/interfaces/IPendingWaitEvent';
 import { CDPSession } from './CDPSession';
 import { NetworkManager } from './NetworkManager';
 import { Keyboard } from './Keyboard';
@@ -199,9 +200,15 @@ export class Page extends TypedEventEmitter<IPuppetPageEvents> implements IPuppe
   }
 
   async close(): Promise<void> {
-    if (this.cdpSession.isConnected()) {
+    if (this.cdpSession.isConnected() && !this.isClosed) {
       // trigger beforeUnload
-      await this.cdpSession.send('Page.close');
+      try {
+        await this.cdpSession.send('Page.close');
+      } catch (err) {
+        if (!err.message.includes('Target closed') && !(err instanceof CanceledPromiseError)) {
+          throw err;
+        }
+      }
     }
     return this.closePromise.promise;
   }
@@ -242,6 +249,7 @@ export class Page extends TypedEventEmitter<IPuppetPageEvents> implements IPuppe
       }),
       this.cdpSession.send('Emulation.setFocusEmulationEnabled', { enabled: true }),
       this.setTimezone(this.browserContext.emulation.timezoneId),
+      this.setLocale(this.browserContext.emulation.locale),
       this.setScreensize(this.browserContext.emulation.viewport),
     ]);
 
@@ -361,11 +369,25 @@ export class Page extends TypedEventEmitter<IPuppetPageEvents> implements IPuppe
   private async setTimezone(timezoneId = '') {
     try {
       await this.cdpSession.send('Emulation.setTimezoneOverride', { timezoneId });
-    } catch (exception) {
-      if (exception.message.includes('Timezone override is already in effect')) return;
-      if (exception.message.includes('Invalid timezone'))
+    } catch (error) {
+      if (error.message.includes('Timezone override is already in effect')) return;
+      if (error.message.includes('Invalid timezone'))
         throw new Error(`Invalid timezone ID: ${timezoneId}`);
-      throw exception;
+      throw error;
+    }
+  }
+
+  private async setLocale(locale = 'en-US') {
+    try {
+      await this.cdpSession.send('Emulation.setLocaleOverride', { locale });
+    } catch (error) {
+      // not installed in Chrome 80
+      if (error.message.includes("'Emulation.setLocaleOverride' wasn't found")) return;
+      // All pages in the same renderer share locale. All such pages belong to the same
+      // context and if locale is overridden for one of them its value is the same as
+      // we are trying to set so it's not a problem.
+      if (error.message.includes('Another locale override is already in effect')) return;
+      throw error;
     }
   }
 }
