+++ puppet/interfaces/IBrowserEmulation.ts
@@ -2,7 +2,7 @@ import IViewport from "@secret-agent/core-interfaces/IViewport";
 
 export default interface IBrowserEmulation {
   userAgent: string;
-  acceptLanguage: string;
+  locale: string;
   platform: string;
   proxyPassword: string;
   viewport: IViewport;
