+++ website/docs/BasicInterfaces/Browser.md
@@ -113,6 +113,9 @@ Update existing configuration settings.
 - options `object` Accepts any of the following:
   - emulatorId `string`. Emulate a specific browser version.
   - humanoidId `string`. Create human-like mouse/keyboard movements.
+  - timezoneId `string`. Overrides the host timezone. A list of valid ids are available at [unicode.org](https://unicode-org.github.io/cldr-staging/charts/37/supplemental/zone_tzid.html)
+  - locale `string`. Overrides the host languages settings (eg, en-US). Locale will affect navigator.language value, Accept-Language request header value as well as number and date formatting rules.
+  - viewport `IViewport`. Sets the browser screen size, window position, inner/outer width and height. If not provided, a random screen, position and viewport will be statistically sampled from data pulled from [statcounter.com](https://gs.statcounter.com/screen-resolution-stats/desktop/united-states-of-america).
   - renderingOptions `string[]`. Controls browser functionality.
   - showReplay `boolean`. Whether or not to show the Replay UI. Can also be set with an env variable: `SA_SHOW_REPLAY=true`.
 
@@ -178,6 +181,7 @@ await newTab.waitForAllContentLoaded();
 ## Aliased Tab Methods
 
 Browser instances have aliases to all top-level Tab methods. They will be routed to the `activeTab`.
+
 ### browser.goBack*()*
 
 Alias for [Tab.goBack](./tab#back)
